requirements.txt
asgiref==3.2.7
bson==0.5.8
dataclasses==0.6
dj-database-url==0.5.0
Django==2.2.12
django-cors-headers==3.2.1
django-filter==2.2.0
djangorestframework==3.11.0
djongo==1.3.2
dnspython==1.16.0
gunicorn==20.0.4
Markdown==3.2.1
mongoengine==0.19.1
pymongo==3.10.1
python-dateutil==2.8.1
pytz==2019.3
six==1.14.0
sqlparse==0.2.4
whitenoise==5.0.1